# backend-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VK Bot API",
        "description": "API для управления VK ботом",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/accounts/all-accounts": {
            "get": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Получить все аккаунты ВКонтакте",
                "description": "Возвращает список всех добавленных аккаунтов ВКонтакте с их основной информацией",
                "operationId": "9e25fb3fe679b8fc20eede1939b3ec7a",
                "responses": {
                    "200": {
                        "description": "Успешное получение списка аккаунтов",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AccountListResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServerErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/accounts/add-account": {
            "post": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Добавить аккаунт ВКонтакте",
                "description": "Устанавливает и сохраняет данные аккаунта ВКонтакте по токену доступа",
                "operationId": "f3da98273182be4dd886528196153af6",
                "requestBody": {
                    "description": "Данные для добавления аккаунта",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "access_token": {
                                        "description": "Токен доступа к API ВКонтакте",
                                        "type": "string",
                                        "example": "vk1.a.abc123def456..."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Данные аккаунта успешно установлены",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Статус успешности запроса",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/AccountModel"
                                        },
                                        "message": {
                                            "description": "Сообщение о результате операции",
                                            "type": "string",
                                            "example": "Аккаунт успешно добавлен"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServerErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/accounts/delete-account/{id}": {
            "delete": {
                "tags": [
                    "Accounts"
                ],
                "summary": "Удалить аккаунт ВКонтакте",
                "description": "Удаляет аккаунт ВКонтакте из системы по его идентификатору",
                "operationId": "31a2a453d0a5148bec71817551db5d5c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Идентификатор аккаунта для удаления",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 123456789
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Аккаунт успешно удален",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Статус успешности запроса",
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "description": "Сообщение о результате операции",
                                            "type": "string",
                                            "example": "Аккаунт успешно удален"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Аккаунт не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "description": "Статус успешности запроса",
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "description": "Сообщение о результате операции",
                                            "type": "string",
                                            "example": "Аккаунт не найден"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServerErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "AccountModel": {
                "description": "Модель аккаунта ВКонтакте",
                "properties": {
                    "account_id": {
                        "description": "Уникальный идентификатор аккаунта ВК",
                        "type": "integer",
                        "example": 123456789
                    },
                    "access_token": {
                        "description": "Токен доступа к API ВКонтакте",
                        "type": "string",
                        "example": "vk1.a.abc123def456..."
                    },
                    "screen_name": {
                        "description": "Короткое имя пользователя",
                        "type": "string",
                        "example": "john_doe",
                        "nullable": true
                    },
                    "first_name": {
                        "description": "Имя пользователя",
                        "type": "string",
                        "example": "Иван",
                        "nullable": true
                    },
                    "last_name": {
                        "description": "Фамилия пользователя",
                        "type": "string",
                        "example": "Иванов",
                        "nullable": true
                    },
                    "birthday_date": {
                        "description": "Дата рождения",
                        "type": "string",
                        "format": "date",
                        "example": "1990-01-01",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "AccountListResponse": {
                "description": "Ответ со списком аккаунтов ВКонтакте",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/BaseSuccessResponse"
                    },
                    {
                        "properties": {
                            "message": {
                                "description": "Сообщение о результате операции",
                                "type": "string",
                                "example": "Список аккаунтов получен"
                            },
                            "data": {
                                "description": "Массив аккаунтов ВКонтакте",
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/AccountModel"
                                }
                            }
                        },
                        "type": "object"
                    }
                ]
            },
            "BaseSuccessResponse": {
                "description": "Базовая схема успешного ответа",
                "properties": {
                    "success": {
                        "description": "Статус успешности запроса",
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "description": "Сообщение о результате операции",
                        "type": "string",
                        "example": "Операция выполнена успешно"
                    }
                },
                "type": "object"
            },
...
            "ServerErrorResponse": {
                "description": "Схема ответа при внутренней ошибке сервера (500)",
                "properties": {
                    "message": {
                        "description": "Сообщение об ошибке",
                        "type": "string",
                        "example": "Внутренняя ошибка сервера при обработке запроса"
                    },
                    "exception": {
                        "description": "Класс исключения",
                        "type": "string",
                        "example": "App\\Exceptions\\ServiceException"
                    },
                    "file": {
                        "description": "Файл, в котором произошла ошибка",
                        "type": "string",
                        "example": "/var/www/html/vendor/package/src/Client.php"
                    },
                    "line": {
                        "description": "Строка, на которой произошла ошибка",
                        "type": "integer",
                        "example": 217
                    },
                    "trace": {
                        "description": "Трассировка стека вызовов",
                        "type": "array",
                        "items": {
                            "properties": {
                                "file": {
                                    "description": "Файл в трассировке",
                                    "type": "string",
                                    "example": "/var/www/html/vendor/package/src/Client.php"
                                },
                                "line": {
                                    "description": "Строка в трассировке",
                                    "type": "integer",
                                    "example": 185
                                },
                                "function": {
                                    "description": "Функция в трассировке",
                                    "type": "string",
                                    "example": "toException"
                                },
                                "class": {
                                    "description": "Класс в трассировке",
                                    "type": "string",
                                    "example": "ATehnix\\VkClient\\Client"
                                },
                                "type": {
                                    "description": "Тип вызова",
                                    "type": "string",
                                    "example": "::"
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}